* Merged new vendor upstream to /vendor  * Merged new upstream code into /trunk
The merge should be valid (4 conflicts, easily resolved by manual merge), but porters should still check.

git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@8 37977431-3df6-0310-b722-df95706aa16b

+++ rts/MoveMath.cpp
@@ -34,8 +34,8 @@ float CMoveMath::SpeedMod(const MoveData& moveData, int xSquare, int zSquare) {
 }
 
 float CMoveMath::SpeedMod(const MoveData& moveData, float3 pos,const float3& moveDir) {
-	int x = pos.x / SQUARE_SIZE;
-	int z = pos.z / SQUARE_SIZE;
+	int x = (int)pos.x / SQUARE_SIZE;
+	int z = (int)pos.z / SQUARE_SIZE;
 	return SpeedMod(moveData, x, z,moveDir);
 }
